pyproject.toml
[project]
name = "nnx-lightning"
version = "0.1.0"
description = "Example of training a Flax NNX model with PyTorch data loaders and the Lightning trainer"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "flax>=0.10.2",
  "jax[cuda12]>=0.5.0",
  "lightning>=2.5.0.post0",
  "pydlpack>=0.2.1",
  "tensorboard>=2.18.0",
  "torch>=2.5.1",
  "torchvision>=0.2.0",
]

[dependency-groups]
dev = ["pytest>=8.3.4"]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu121", marker = "platform_system != 'Darwin'" }]
torchvision = [
  { index = "pytorch-cu121", marker = "platform_system != 'Darwin'" },
]
